Changed function paramter type: uint16 ->uint8

+++ Byggern/Byggern/src/drivers/latch.h
@@ -12,7 +12,7 @@
 #include <asf.h>
 
 void latch_init(void);
-void latch_write(uint16_t val);
+void latch_write(uint8_t val);
 
 
 #endif /* LATCH_H_ */
